renamed 'mips_cpu_impl' -> 'mips_cpu'
